generalize ExecValidator method
